debugger extensions are removed only in prod environment

+++ src/opt/docker/provision/entrypoint.d/11-php-debugger.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
 
-for CONF_PATH in ${PHP_CONF_PATHS}; do
-    rm -f "${CONF_PATH}"/xdebug.ini
-    rm -f "${CONF_PATH}"/blackfire.ini
-done
+if [[ ${DOCKER_XAMPP_BIN_ENV} == "prod" ]]; then
+    for CONF_PATH in ${PHP_CONF_PATHS}; do
+        rm -f "${CONF_PATH}"/xdebug.ini
+        rm -f "${CONF_PATH}"/blackfire.ini
+    done
+fi
